IMAGE-specific classes moved to separate package
