+++ app/src/main/java/com/example/hw/MainActivity.java
@@ -55,6 +55,10 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
+import com.image.MakeRequest;
+import com.image.MapRequestFormat;
+import com.image.PhotoRequestFormat;
+import com.image.ResponseFormat;
 import com.scand.svg.SVGHelper;
 
 import org.json.JSONException;
